requirements.txt
# === Requirements ===
# Dependências do projeto ABI
# Execute: pip install -r requirements.txt

# === Framework Web ===
streamlit==1.28.1

# === IA e LLM ===
google-generativeai==0.3.0

# === Utilities ===
python-dotenv==1.0.0

# === Desenvolvimento ===
pytest==7.4.3
black==23.12.0
flake8==6.1.0
mypy==1.7.1
